Change int to float

+++ calc.py
@@ -1,8 +1,8 @@
 def compute(expression):
     values = expression.split(' ')
-    num0 = int(values[0])
+    num0 = float(values[0])
     operator = values[1]
-    num1 = int(values[2])
+    num1 = float(values[2])
     if operator == '+':
         return num0 + num1
     else:
